HW-1 리포지터리 패턴 적용 , 카테고리 정렬 로직 개선
