+++ dlwls5201/MyArchitecture/app/src/main/java/study/architecture/myarchitecture/repository/UpbitRepository.kt
@@ -1,0 +1,12 @@
+package study.architecture.myarchitecture.repository
+
+import io.reactivex.Single
+import study.architecture.myarchitecture.network.model.UpbitMarket
+import study.architecture.myarchitecture.network.model.UpbitTicker
+
+interface UpbitRepository {
+
+    fun getGroupedMarkets(): Single<Map<String, List<UpbitMarket>>>
+
+    fun getTickers(markets: String): Single<List<UpbitTicker>>
+}
